Non-ag cells masked in Kc calculation
I'm not exactly sure where to apply the masking, but doing it in the Kc calculation seemed like the right spot.  I'm not going to close the issue (#4) yet since I think it would be good to be able to control the masking via an Image class parameter flag and we should probably decide if any of the other bands should be masked (ETr is the main one).

Updated all the example notebooks to show the masking.