Updated README and CONTRIBUTING
The README still needs a small summary explaining how the SIMS model works.  It would also be good to add some references, at a minimum we should pull the Johnson and Trout 2012 reference from the Image class.

We may want to drop all the badges until the repo is public.