Working through some ideas for restructuring
I think the crop_kc data should be stored in one object for all crop_types instead of as a function of crop_type.  This would make it much easier to use this data in the model.py code instead of having additional lists of field and tree crops and mappings from crop_type to crop_class.