Added basic logic was setting water Kc
We will probably want to add more logic for determining water pixels (NDVI < 0 is a little simplistic).  We should also decide if the water Kc override should be done before or after the other crops and if it should apply to crop pixels (currently it is only non-ag crops that can be water).

Renamed "crop_type_mask_flag" to "mask_non_ag_flag" to make it more obvious what the flag does.