Trying out combining the aggregation codes into a single function
This makes it easier to control the output type the same way across all the time intervals.

These blocks of code were almost identical so combining them seemed like a good idea.  The one problem is that the mapped function can only take one input, so it is still necessary to build separate functions for each aggregation time step but then have them call a common aggregation function.  It probably isn't ideal to have the daily function use this since it makes an extra .sum() but this may not be a big deal.

Changed the default output type back float.

Shortened the daily interpolation test time period since a full month was raising a memory error.