ETr parameters were not being applied correctly/consistently in Collection class
In general it might be better if the Collection class and methods didn't take the etr parameters as separate parameters and instead read them from the model_args or the kwargs.  As they are now, the values need to be set back into model_args (in interpolate) so that they can get read inside the build and overpass methods.

Also, changed the default etr_factor to None for the Collection class so that it must be set by the user if computing ET or ETr (similar to etr_source and etr_band).  I would kind of like to have these be mandatory parameters instead but then you can't use the interpolator for just ETf or NDVI calculations.